+++ packages/app-extensions/src/appFactory/appFactory.js
@@ -1,10 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import _union from 'lodash/union'
-import {consoleLogger} from 'tocco-util'
+import {consoleLogger, intl} from 'tocco-util'
 
 import errorLogging from '../errorLogging'
-import intl from '../intl'
 import App from './App'
 
 export const inputDispatchActionType = 'app/INPUT_DISPATCHED'
