+++ packages/app-extensions/src/cache/index.js
@@ -1,22 +1,7 @@
 import {
-  addShortTerm,
-  addLongTerm,
-  getShortTerm,
-  getLongTerm,
-  removeShortTerm,
-  removeLongTerm,
-  clearShortTerm,
-  clearAll,
   addToStore
 } from './cache'
+
 export default {
-  addShortTerm,
-  addLongTerm,
-  getShortTerm,
-  getLongTerm,
-  removeShortTerm,
-  removeLongTerm,
-  clearShortTerm,
-  clearAll,
   addToStore
 }
