+++ packages/app-extensions/src/cache/sagas.js
@@ -1,7 +1,7 @@
 import {all, call} from 'redux-saga/effects'
+import {cache} from 'tocco-util'
 
 import rest from '../rest'
-import {clearAll} from './cache'
 
 export default function* mainSagas() {
   yield all([
@@ -15,7 +15,7 @@ export function* init() {
   if (!clearCacheChecked) {
     const revisionChanged = yield call(rest.hasRevisionIdChanged)
     if (revisionChanged) {
-      yield call(clearAll)
+      yield call(cache.clearAll)
     }
     clearCacheChecked = true
   }
