+++ packages/app-extensions/src/rest/helpers.spec.js
@@ -3,10 +3,10 @@ import {expectSaga} from 'redux-saga-test-plan'
 import * as matchers from 'redux-saga-test-plan/matchers'
 import {call} from 'redux-saga/effects'
 import {throwError} from 'redux-saga-test-plan/providers'
+import {cache} from 'tocco-util'
 
 import {requestSaga} from './rest'
 import * as helpers from './helpers'
-import cache from '../cache'
 
 describe('app-extensions', () => {
   describe('rest', () => {
