+++ packages/login/src/modules/sagas.js
@@ -1,5 +1,5 @@
-import {consoleLogger, request} from 'tocco-util'
-import {cache, externalEvents, intl, rest} from 'tocco-app-extensions'
+import {consoleLogger, request, cache, intl} from 'tocco-util'
+import {externalEvents, rest} from 'tocco-app-extensions'
 import {takeLatest, put, select, call, all} from 'redux-saga/effects'
 
 import * as actions from './actions'
