+++ packages/tocco-util/src/cache/index.js
@@ -1,0 +1,20 @@
+import {
+  addShortTerm,
+  addLongTerm,
+  getShortTerm,
+  getLongTerm,
+  removeShortTerm,
+  removeLongTerm,
+  clearShortTerm,
+  clearAll
+} from './cache'
+export default {
+  addShortTerm,
+  addLongTerm,
+  getShortTerm,
+  getLongTerm,
+  removeShortTerm,
+  removeLongTerm,
+  clearShortTerm,
+  clearAll
+}
